+++ lib/browser/device/playstation3.rb
@@ -12,7 +12,7 @@ module Browser
       end
 
       def match?
-        ua =~ /PLAYSTATION 3/i
+        browser.ua =~ /PLAYSTATION 3/i
       end
     end
   end